added authentication restriction to delete and update room

+++ base/views.py
@@ -8,6 +8,7 @@ from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login,logout
 from django.contrib import messages
 
+from django.contrib.auth.decorators import login_required
 
 
 def loginPage(request):
@@ -55,6 +56,8 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'base/room.html',context)
 
+
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -65,10 +68,14 @@ def createRoom(request):
     context ={'form':form}
     return render(request,'base/room_form.html',context)
 
+
+@login_required(login_url='/login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user!=room.host:
+        return HttpResponse('You are not allowed to update this room!')
 
     if request.method=='POST':
         form = RoomForm(request.POST,instance=room)
@@ -79,8 +86,14 @@ def updateRoom(request,pk):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+
+@login_required(login_url='/login')
 def deleteRoom(request,pk):
     room =Room.objects.get(id=pk)
+
+    if request.user!=room.host:
+        return HttpResponse('You are not allowed to delete this room!')
+
     if request.method=='POST':
         room.delete()
         return redirect('home')
